pyproject.toml
[tool.poetry]
name = "kaggle-tps-mar-2022-odsc"
version = "0.1.0"
description = "Using PyCaret, W&B, and Prefect to show how to orchestrate reproducible AutoML experiments on the Tabular Playground Series - March 2022 on Kaggle"
authors = ["Anish Shah <anish@wandb.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
kaggle = "^1.5.12"
wandb = "^0.12.11"
numpy = "^1.22.3"
pandas = "^1.4.1"
prefect = {extras = ["viz"], version = "^1.0.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyterlab = "^3.3.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
